backend/requirements.txt
fastapi
langchain
numpy
openai
pandas
pinecone-client
pydantic
requests
SQLAlchemy
huggingface_hub
python-dotenv
cohere
supabase
pymongo
langchain_community
